Basic skeleton to check RTL-SDR handshake
I am in the process of moving my old FFT Spectrum Analyser logic from
Octave to Python based one. Also initialy checking it against rtl-sdr.

It gets a handle to rtl-sdr connected to the computer.

Followed by
setting the sampling rate at which data has to be got.
The center frequency is set wrt the tuner.
The gain is set 0.

Then few samples are read from the device. I assume the complex
data got from the device is I data as the real part and Q data as
imaginary part of the complex number.

Also the details of the sdr handle is printed using dir(sdr_handle)