Basic Reorganisation of Code Part3 Plot related
The internal plot logic which uses Cairo library has been reorganised
a bit, as well as the functions renamed to clearly indicate it uses
cairo library for drawing routines.

The Y Axis auto dimension logic, has been moved out of the cairo
plot_setup logic to the new cairoplot_data function created to plot a
set of data. This is because the Axis dimensions can be got only from
the data to be plotted.

Also the Cairo and MatPlotLib related plot has been moved into a
common plot_data function.