plot_data dB finetuning P2
Have changed the smallest measurement used as reference from
1/255 to 2/256.

Have clipped fft results which are smaller than 25% of the ref
value considered for dB calculations. This is to avoid the useless
large negative dB values in the plot, which will skew the scale
used for plot that is displayed along the Y-axis.

Also added a new iDEBUG ==4 block, which allows one to see the
fft results directly, as well as after converting to dB levels.