SCAN Live Plotting
Added support to do live plotting of the fft results as the logic
is sweeping thro the scan range specified by the user. Previously
it would have updated only after the full sweep is finished.

This is mainly achieved by using the matplotlib.pyplots show func
with block=False argument, asWellAs the pause(someFractionOfASecond)
function.

Also pyplot cla is required to be called before a new dataset is
plotted, so that the plot library or its gui toolkit doesn't get
overloaded with previous plot data. Without this as more and more
datasets are plotted, it takes more time to plot and thus slows
down the program.