ZeroSpan Live Repeat logic
Added new logic which works in zero span mode, i.e it doesn't sweep
from one frequency band to another (band here is determined by the
sampling rate specified by us). Instead it keeps scanning the same
band again and again. And inturn it averages the fft results as
well as plots the same.

Moved the bLivePlot to the beginning of the file and renamed it has
gbLivePlot. Inturn the main logic now checks if gbLivePlot is enabled
or not to decide whether this new zerospan_repeat logic is called or
if the original one time zerospan plot logic is called.

NOTE: As of now this new zerospan_repeat doesn't bother to convert
the fft result to dB and inturn adjust wrt the gain set wrt rtl-sdr.