ZeroSpan_Repeat logic YAxis Limit update
Added a new global variable and corresponding logic, which if
enabled will allow the live plot in the zerospan_repeat mode to
adapt its min and max Y limits to the fft results (i.e the
amplitude/power in the given frequencies) as it changes. Note
that this is different from the default autoscale logic for the
plots. In the default autoscale logic the min and max keep
increasing or decreasing depending on the current data to be
displayed. While in this AdapativeFixedYLimits logic the Min
and Max are adjusted only towards their respective ends based
on the data from the fft. i.e the MinLimit will never increase
it can only decrease, while a MaxLimit can only increase in its
value and not decrease.