exp3_fft.py
Experiment with sine waves, fft, sampling rate, number of samples
considered etc.

It is to be noted that if the number of time domain samples used
for fft is not such that sample rate is a integer multiple of
the number of samples used, then the power level for the freqs
will not be properly obtained.