exp3_fft.py captureTime and NonMultiple freq wrt samplingRate
Even when the samples are captured in the time domain from some
where inbetween and not from exact beginning, the fft still maintains
its behaviour, as noticed before.

Even having signals whose freq is such that the samplingRate is not
a exact integer multiple of it, doesn't impact the fft results.

Still noticed that if the numberOfSamples in the time domain which
is used for fft analysis, is such that the samplingRate is a integer
multiple of it, then the fft results across frequencies is balanced
and doesn't give any undue advantage to one or the other freq.