exp3_fft.py - dataSamples used for Fft, saving plot p2
Changes
--------
As the loop count in exp_fft is used to decide the length of the data samples
to be processed for fft at a given time, I have made it into a global variable.

And inturn I calculate the Number of Rows to plot as well as the size of the
plot canvas required from this now. So easier to check over a wide range easily
now with out having to worry about adjusting the num of rows in plot and plot
canvas size manually.

Now I plot the data sample being used for fft analysis first before plotting
the fft results. So adjusted the subplot related index calculation suitably.

Also I have changed from ps to pdf format for the saved plot file. Because
ps also depends on the papertype in addition to the figsize set in figure
command. So with ps, even if the figsize is set properly (which helps spread
the subplots properly in the plot), still the saved plot could contain a
truncated plot, as the papertype to savefig decides the saved plots size.
However with pdf, as the figsize is changed in figure, the generated pdf files
dimensions are adjusted suitably.

Observation
------------

Now I have noticed few discrepencies wrt my earlier conclusions.

1) Even thou 1250 is a perfect submultiple of 10000, still the fft result is
not proper, but is skewed. (This is the first occurance of skew for a
submultiple in this experiment).

2) When using 3000 samples for fft, the results seem balanced. However when
using 3333 samples for fft, the results seem unbalanced. (This continues the
earlier noticed skew that for certain rational fractions of the samplingRate,
the fft result is skewed and not balanced)