Oct nfft nyqfreq, python samples-Fft test changes
Add forgotten to commit the octave normalised fft function change
to limit its returned values to +ve freq's ie upto Nyquist freq,
from a week or so back. Its commited to git now.

For the python test code, which brute force checks the impact of the
samples selected (interms of as to from when and how much) on the
simple fft analysis output,  has been updated to

* now generate a test signal which contains a bunch of equally spaced
  test frequencies in it.

* if the specified fft bin size is smaller than the sampling rate,
  then to window-slide through the full data set containing one full
  sample rate worth of data in a non-overlapping way, [Note: the full
  data set starts from a randomly predetermined starting phase for
  the signals], such that the calculated fft of each window (of fft
  bin size) is cumulated or rather averaged over the non overlapping
  window slide till the end of samples

and inturn check how fft output is impacted. In the previous run/version
of the test logic, one was only looking at one bucket of sample data
at the begining of the set of samples, when the bucket/bin size was
smaller than sampling rate.

This test logic tries to identify the impact of sample selection on
fft output, by discarding some signal data at the begining, which is
determined randomly, so that the starting phase of the signal is
different each time the test program is run, as well as by progressively
checking with smaller and smaller set of data from the generated sample
data and compare the output results.

NOTE: This is a post facto commit after more than 3 years, as usual,
due to me becoming busy with some other stuff at that time, I had
abondened this brute force test experiment at that time. The above is
what I gather now after a very very quick glance.

Have added the helper run script I had created for reference wrt using
the specanal, also into git now.