PartialCycleWindowedFft: Conv to RectWindow equivalent
Use the ratio of the rectangle window area and the area of window
function of interest as the conversion factor for the value got by
fft of windowed sample to the equivalent value if rectangle window
had been used.

Add grid to the fft plot.

Also add a 3rd freq to the test sample.