Exp4: Cleanup FftMax's minmax tracking, area of sig
NOTE: Min is set to a high value and Max is set to a low value,
so that as the logic runs, the values can get updated as required,
without lossing anything.

TODO: Currently this logic doesnt handle multiple frequencies with
in the test signal, properly fully. But still at a different level
still gives the required info to a great extent roughly.

Also added logic to do_it to optionally calculate the area under
the amplitude curve in time domain. It also shows that the value
is the same across the different frequencies.

This along with the plots, also highlights the issue with a direct
fft of the signal (i.e a rectangle window) not depicting all signals
with same amplitude, in a equivalent manner when transformed into
the freq domain. At same time the windows like Kaiser, Hann, ...
make the different freqs in the signal to appear relatively similar
in freq domain, as they are in the time domain. And by compensating
for the lost area/amount wrt windowing, we can get better representative
data.