Exp4: Less data and Overlapped sliding
With less data to work with in a given fft run, sliding with overlapping
over a larger data set, still doesnt help recover the true nature of the
signal in the full / larger data set.

NOTE: Logically seems to make sense, but need to refresh the theorey and
look at any other possibilities later.