Exp4: Overlapped Sliding good for Rect Window also
At a simple level and without digging into anything, the learning
could be

If one has only 1sec and or bit more of time domain data, then
using a smoothing window function on the data before fft, gives
the better fft result, which is representative of the signals in
the sampled data.

If one has few seconds and or bit more of the time domain data,
then a overlapped sliding data set of around 1 sec and or bit more
with direct fft (i.e rect window and not anything fancy ones like
kaiser or hann or ...) along with averaging of the cummulated fft
data, gives a very good representation of the time domain signals
even in the frequency domain.

kspecanal has a equivalent note, for people who may want to use it.
Rather Exp4 confirms the note.