simple kspecanal - overlapped sliding curscan
As the idea behind this program was to help identify any emissions
and then tweaking designs in a pinch, where relative difference
btw measurements is good enough, at one level. So once I was
able to achieve the same with the basic skeletal flow, I had
stopped mid way through the logic.

On thinking further, as being able to identify the frequencies
involved is better than playing hide and seek with dwell time,
at a basic level, so for now abondoning the initial dwell time
and related path with partial windows etc, which either way adds
complexities, without gaining much, and instead going for a full
on few seconds based scanning.

Logic changed to work with full sample rate size window at a time,
along with the skeleton to do overlapped sliding.