kspecanal: reading large datasets, float to ints
As the new logic requires processing multiples of sample rate size
sample buffers, so one needs to read these large sets of data from
rtlsdr. Inturn trying to read large size buffers from rtlsdr is
triggering libusb overflow errors. So have changed the logic to
read the required size of buffer over multiple runs, from the device.

As it was documented to try and read buffers which are powers of
2 from the device, so have updated the logic to account the same.

So also as the logic now does overlapped sliding window based fft
processing, so I round up to the next nearest power of 2 buffer
size, which can be implicitly handled by the sliding logic.

At the same time also take care of exiting the sliding loop, if
the amount of data is not sufficient, towards the end of looping.