kspecanal: Changed the defaults a bit
For the default ZERO_SPAN mode, now the gain is increased to 7.1
instead of the default 0, so that user can see the signal. Same
for SCAN mode. However for FULL_SCAN mode, the gain has reduced
from 48 to 7.1

Similarly the default sampling rate and center freq is also set
through global variable now.