kspecanal: readsamples, gNonOverlap, -1 fftsize
As it is noted that read_buf data output is potentially from a
shared buffer, which could change values, and also because read_sample
does the normalising of IQ data and also puts it in complex number
format, so also triggering fft on true amplitude rather than the
mixed i and q wrt my logic, so for now have changed to using read_sample.

The percentage of overlap is controlled through a global variable now.

The gFftSize is set to -1, so that it gets set to sampling rate by
default. But user can change it, if required.

Also added nextpow2, as its missing in numpy.