curscan: extract required fft result subset, after discarding dc
The logic, like previous version, returns the DC component as a
seperate member in the return set.

NOTE: fftsize is assumed to be even, and the freq component which
is shared between the +ve and -ve frequencies is also returned now.

As the fft result plot was showing some spurious lines, I did these
changes so that the logic is fine (by avoiding the known things) and
ensure that the spurious line is not bcas of dc (shouldnt have been
as I was setting it to a safe noise floor) or so, just to be doubly
sure. And as was expecting the spurious peaks/lines still exist.