Fix some oversights and complex dft dejavu
With the last few changes including switching to direct complex
representation of the signal rather than the force mixed real,
which I used (rather blindly, now come to think of it, instead
of trying to think a bit more before jumping) before, I was
reviewing the flow in general. Also as part of same, I stumbled
on the centered or not flow which I had put as a quick fix
originally.

And thats when the things fell into frame and the missteps hit
me.

How on earth could I bliss fully forget complex dfts and its
implications, the -ve freqs, oh god/freak, and make a bit of
mess of things unnecessarily at one level for certain aspect ;-)
Am i discarding half of the data and introducing a forced kind
of noise, by using the data wrongly. Need to explore this a bit.

Also fix a oversight wrt moving around the complex signal.

Also revert to allowing the zero span plot window to adapt freely
wrt the data available, rather than forcing it to only zoom out,
but not zoom in, as and when required.