SpecAnal: _scanRange setup for cumulate
Now scan_range and _scan_range have the corresponding full range
Freqs and Ffts arrays which they pass around, so that the fft results
from one round of scan range can be passed on to the subsequent rounds.
This will be used for cummulation with Max or Avg logic later.

If no array is passed to _scan_range, then it creates one.

If there are any Infinities as a result of log conversion, then the
same is reset to 0, so that the entry is plotted with y pos of 0
for amplitude.