Quit on unknown cumuMode, compare upper case
The user entered cumulation mode is converted to upper case and the
logic inturn checks against mode in upper case. So user can enter
mode either in upper or lower or mixed, it doesnt matter.

Also if the user entered mode is not known, then the program will
alert the user and quit.