specanal: imshow.set_data, autoscale
To avoid overhead and spead up things and be efficient, now the
logic creates an instance of imshow and inturn keeps updating the
data set associated with it, in the loop. This ensures that the
logic is relatively efficient and fast.

Also had to use autoscale, so that the data set wrt imshow can be
interpreted and color mapped properly. Otherwise, when imshow is
called at the beginning the data to color mapping boundries are
determined with the data values at that time, which is all zeros
in this case; and then updating the data later using set_data,
doesnt update this mapping boundries, so imshow window content dont
get updated visually, as all non zero data is ignored wrt color
mapping.

NOTE: Need to think, whether to leave this at autscale or explicitly
set the vmin and vmax, or maybe better yet provide option to user,
rather than deciding for them one way or the other.

Also added a helper script to show some possible ways of calling
the program.