sdr_read large buffer in steps, mult of 2**16
If the amount of data to read from rtlsdr is larger than 64K, then
it is read in steps of 64K, to avoid possible libusb io errors.
This also means the fullSize should be a multiple of 64K, if it is
larger than 64K.

NOTE: // used instead of / to get a integer division.

The way fullSize is calculated from fftSize also updated, such that
for relatively large fftSize, the fullSize multiplier is smaller,
while also allowing the full gamet of 2**N size buffers to be
worked with.