SpecAnal:ScanRange: Ensure partial scan does go through
When doing a overlapped sliding based stepping through the freq bands
as part of covering the full freq range specified for scan range, one
may require to do a partial freq band scan wrt the last step.

This is handled now. Previously it would have skipped the last step/
scan, if it was a partial scan.