SpecAnal: Try handle sdr setup failure - p1
When trying to set the center freq to anything between 949 - 949.9MHz,
the rtlsdr with FC0013 is failing. So tried to trap the same and
skip to next freq band in _scanRange.

But when trying to set the next freq band(which is valid on its own),
after the pll failure related to 949-949.99, it is segfaulting.

TODO: Have to try trap in sdr_setup itself and then close and reopen
the sdr device to see, if things will be fine.