SpecAnal: Try handle sdr setup failure P2
In sdr setup, on failure, close and reopen a new sdr instance.
The same is returned to the caller.

To help ensure that the new sdr instance is what is used beyond
that time, now the sdr is added into the dictionary passed around
by the logics in the program.