Cleanup: clearInfs, setup discard,
When converting to log10 as part of data_proc, if there are 0s, then
it will get set to Inf. Now the option to reset these to a predefined
value is handled within data_proc itself as a optional logic.

sdr_setup will call the read_samples to discard some data at the
beginning, only if the setup was successfull.