SpecAnal: scan_range levels plot default to floor
Previously the scan_range initial default value for fft results
was set to appear at the top by setting to 0. Now it is changed
to set it to the minAmp4Clip value, so that it appears as the
floor in the plot by default.

As samples are read and fft applied, actual value is updated into
the result buffer.

NOTE: However the HeatMap related fft result buffer is still retained
at 0, as that will appear as the base with no signal. While 0 to
log10 wrt Levels plot, forces these to -Inf, which is inturn reset
back to 0 there.