SpecAnal: data_plotcompress - raw, avg, max
This reduces the data resolution so that it fits a predefined graphics
resolution along the x axis. To achieve the same it breaks the data into
a set of adjacent groups and then it collates adjacent data into single
data by either selecting the max value in the group or average of values
in the group.

The user can also decide to ignore this and ask it to return the input
data as such, without any processing, by setting it to raw.