SpecAnal: ZeroSpan updated wrt single figure based flow
NOTE: aspect set to auto instead of the default equal, so that imshow
occupies the full space given to it in its axes/subplot. This makes
the pixels non square, but in our case it si good, because one can
see the frequencies and their level wrt heatmap more clearly.