SpecAnal: Update Layout, show labels wrt xticks wrt axes
Repositioned the layout, such that buttons axes is moved to the
right of heatmap. This also ensures that heatmap and sig levels
subplots are below one another and almost equal size.

Changed figure to use constrained layout so that the tick labels
can be seen clearly.

Updated logic so that xticks and xtick_labels are set seperately,
as required when working with a axes handle.