SpecAnal: GUI/Figure cleanup
Added buttons for different actions. TODO: Actual action needs to
be implemented yet.

The AxFreqs puts the list of high siglevel freqs as texts.

Clear siglevel plot, as the freq bands is being stepped during scan,
ie go back to the original logic wrt this.

Added a window title to the figure window.

Avoid ticks for the AxFreqs subplot/axes.