SpecAnal: Added FMScan and Moved QuickFullScan
Also the logic no longer waits for user input, after adjusting the
scan range, if it is not a multiple of sampling rate.

Also changed the default prgMode from zerospan (around a specific
fm freq of around 92e6 or so) to fmscan.

Added a xlabel for AxFreqs so that user knows what it is.