Test SpecAnal fft and related logic - P1
Added a simple logic to generate signals of few freqs around
specified center frequency, for the specified duration|size
with the given sampling rate.

The signal is offset in time by a random amount, so adjacent
requests for reading sample will be satisfied with samples which
are disjoint in time.

Currently it returns just real data rather than complex iq data.