SpecAnal: curScanCumuMode user property added
Now user can control how curScan cumulates signal level data, as it
slides through the fullSize data set and applies fft in a overlapping
sliding window of fftSize way.

By default it is set to Avg, which is good enough in general. However
there is also the raw and max options, just in case, if one requires.