TestFFT: Show impact of windowing
Show impact of windowing on fft results.

However the loss in signal level due to windowing is not compensated
in this.

Issue of fft result bleading from one end to the other end, if there
is signal at either end of the freq band can be seen by running with
center freq of 11.2e6 or 14.8e6