SpecAnal: pltCompressHM for heatmaps
Have updated the heatmap related logics of both zerospan and scanrange
to use the property pltCompressHM to decide, what data to pass to the
heatmap.

Inturn the same has been set to Max for now. This matchs the previous
gb2DPltCompressForceMax mechanism, which has been removed now. Also the
old logic was only implemented for SR and not ZS.

TODO: Need to avoid PltCompress related recalculating of all rows of
HeatMap data in SR logic, each time.