SpecAnal: Add Min levels also to the mix
Now both cumuModes and pltCompress support Min option, which stores
the minimum of the values seen wrt each signal freq bin, till then.

Inturn ZeroSpan and Scan(Range) now also show min as part of the
siglevels plot.

NOTE: ZeroSpan sets the default to the 1st scan results, so nothing
specific to handle wrt min or max or ... wrt initial value. While
ScanRange starts out with a predefined initialisation value, so for
the case of Min it is set to a relatively large value, so that it
can settle to the actually scanned min values, as they become available.