SpecAnal:UI: Added buttons to toggle Min/Max/Avg siglvl plots
These buttons can be used to toggle the bDataMin, bDataMax and
bDataAvg properties. Which inturn control whether the corresponding
data buckets are created,processed and plotted or not.

The properties are also added to the global dictionary now.

NOTE: There is no seperate control for Cur data bucket, because this
is unique only in zerospan logic, and there the plot should be spacious
enough to show current lvls in addition to others. While in scanrange
logic, the cur also relates to avg in a way ideally, so need to think
whether to maintain it seperately or not.