SpecAnal: Atleast 1 SigLvls Curve On, Btn UI Seq
If user disables all the signal levels curves, then the logic will
automatically enable the Avgs curve.

The logic prevously had been updated such that it uses the last
pltCompress processed data for showing the high sig levels. In turn
the curves are processed in a specific order such that if multiple
curves are active, then the high siglevel markers are applied to a
specific curve. This priority order is (Low)Curs->Avg->Min->Max (High).

The buttons are also sequenced in the GUI in a equivalent order.