SpecAnal: Show freq corresponding to location clicked on heatmap
Added a pick_event handler for the figure and inturn enabled picker
for imshow, so that when user clicks on the heatmap, the logic can
show the freq corresponding to the location clicked on the heatmap.

NOTE: This is needed, because once one sets the xticks for a axes,
the default formatter which maps the location on the axes to the
data space, gets cleared; so this additional logic is required.

Also Soul is a soulfull movie ;-)