SpecAnal: Click location on heatmap to see its freq in xlabel
Instead of showing the freq of location clicked on heatmap, as part
of the heatmap data space, it is now shown as part of the heatmap's
xlabel. This makes the logic simpler, as well as ensures that a dark
heatmap doesnt make the freq printed difficult to read.

Readme updated to indicate this freq identification behaviour wrt HM.