SpecAnal: fftSize < xRes, quickfix kiss soln for scanRange
Rather if fftSize is smaller than xRes, then change xRes to match
the fftSize.

Wrt scanRange, this will lead to loss of resolution if using pltCompress
of Max or Min or Avg, wrt the info got from fft. But if pltCompress raw
or conv is used, then the program will work as well as no info is lost.
So using this for now as a quickfix. INTURN the quickfullscan updated to
use this new semantic, as total number of entries when using pltCompress
raw will work fine for plot, without slowing things down.

In case of zerospan eitherway from fft we wont have greater than fftSize
number of bins, so changing xRes to match fftSize doesnt lead to
any loss of info.  So also with this some of the changes done in the
last commit, is technically not required, but either way keeping it
on still, as those changes are harmless in a way.