SpecAnal: Cross check logic against Pyplot PSD : P1
Modified the logic to use pyplot's psd, instead of my logic.

Initial check seems to indicate, my logic and psd both give similar
result. Also PSD by default seems to use hanning window, similar
to what I was using for the window func previously, when I was not
allowing user to select which window func to use.

As psd plots to the currently active axes, which is the quit btn
related axes, so also I call cla to clear the same, so that psd's
ylabel doesnt mess with the layout. So also one cant see quit.