SpecAnal: Find a suitable xRes - P1
If xRes is not a multiple of fftSize, then find a suitable xRes

This tries to find a suitable xRes which is a multiple of fftSize,
by trying to find from the largest xRes towards the smallest xRes.

TODO: Do it other way round, as we want to find the smallest yet
suitable xRes preferably as default by program.

InterestingMovie: Backstabbing for beginners