SpecAnal: sdr_read handle length < gSdrReadUnit wrt new flow
In a previous commit sdr_read had been updated to handle buffer
length requests which are not a multiple of gSdrReadUnit.

However at that time, it was handled only for the case, where the
length is also larger than gSdrReadUnit.

Now it has been updated to work for requests which are smaller than
gSdrReadUnit, also.

Also with this, revert the default gSdrReadUnit back to the larger
value, as the logic will handle smaller values appropriately now.