SpecAnal: _scanRange Fft.Cur is base, Dont go beyond reqd range
As Fft.Cur is no longer a raw scan value, but average of the scans
at that freq during the current full scan's sliding, and inturn
used as the reference for the other curves like Min/Max/Avg, so
it needs to be calculated always, irrespective of whether user
wants cur curve to be available or not.

Ensure that we dont go beyond the required freq range, when saving
into the curves.