SpecAnal:_scanRange ensure Fft.Avg calcd for Heatmap
Fft.Avg is calculated irrespective of if user requires Avg curve
to be displayed, because it is also used by heatmap logic. This
was the case originally, however when the logic was changed in prev
commits wrt using avg cur of cur full scan for Max/Min/Avg, this
had got changed, this is fixed now.